refactor: 캐릭터 API 패키지 구조 정리
