+++ src/main/kotlin/com/team2/server/party/controller/CharacterApi.kt
@@ -1,8 +1,8 @@
-package com.team2.server.party.api
+package com.team2.server.party.controller
 
 import com.team2.server.common.response.ApiResponse
 import com.team2.server.common.swagger.InternalServerErrorResponse
-import com.team2.server.party.api.dto.CharacterResponse
+import com.team2.server.party.dto.CharacterResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import io.swagger.v3.oas.annotations.responses.ApiResponse as SwaggerApiResponse
