+++ src/main/kotlin/com/team2/server/party/dto/CharacterResponse.kt
@@ -1,6 +1,5 @@
-package com.team2.server.party.api.dto
+package com.team2.server.party.dto
 
-import com.team2.server.party.application.dto.CharacterResult
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "캐릭터 조회 응답")
