+++ src/main/kotlin/com/team2/server/party/dto/CharacterResult.kt
@@ -1,4 +1,4 @@
-package com.team2.server.party.application.dto
+package com.team2.server.party.dto
 
 data class CharacterResult(
     val characterId: Long,
