+++ src/main/kotlin/com/team2/server/party/usecase/GetCharactersUseCase.kt
@@ -1,8 +1,8 @@
-package com.team2.server.party.application.usecase
+package com.team2.server.party.usecase
 
 import com.team2.server.common.entity.ImageTargetType
 import com.team2.server.common.repository.ImageRepository
-import com.team2.server.party.application.dto.CharacterResult
+import com.team2.server.party.dto.CharacterResult
 import com.team2.server.party.repository.CharacterRepository
 import org.springframework.data.domain.Sort
 import org.springframework.stereotype.Service
